Re-adding errantly removed graphics
When a new graphic's info window is displayed and then the map is
clicked, the graphic disappears. It seems that the ArcGIS API for
JavaScript is removing those graphics. I don't know how to stop it, but
with this commit I'm listening for the graphic-remove event and
re-adding the graphic.

Fixes #10